+++ frontend/src/components/CicloRegister/novoRegistro.ts
