--- frontend/src/components/CicloTable/cicloRegistro.ts
@@ -1,8 +1,0 @@
-export type cicloItem = {
-    id: number,
-    materia: string,
-    horasPausa: string,
-    horasFeitas: string,
-    progresso: number,
-    tags: string,
-}
